fix(codec): drop redundant `u16::MAX as u16` cast + needless borrow
- state_machine.rs:279 — `(u16::MAX as u16).to_le_bytes()` tripped
  clippy::unnecessary_cast under CI's `-D warnings`. The cast is a
  no-op since `u16::MAX` is already `u16`; use `u16::MAX.to_le_bytes()`
  directly.
- examples/sm_fixtures.rs:77 — `bytes_to_hex(&bytes)` tripped
  clippy::needless_borrow because `enc.finish()` returns `&[u8]`, so
  passing `&bytes` re-borrows the slice. Drop the leading `&`.

`cargo clippy -p ciri-protocol --all-targets -- -D warnings` now clean.
The two remaining workspace-wide errors (`ciri-motion`
clippy::manual_range_contains, `ciri-config/src/config.rs`
clippy::type_complexity) sit in files this branch does not touch and
are pre-existing on `dev` — out of scope for the web worktree.

+++ crates/ciri-protocol/src/codec/state_machine.rs
@@ -275,8 +275,7 @@ impl StateEncoder {
                     // wrapped count's worth of cells, mis-parsing the
                     // rest of the stream as garbage opcodes.
                     self.out.push(OP_CHARS_LONG);
-                    self.out
-                        .extend_from_slice(&(u16::MAX as u16).to_le_bytes());
+                    self.out.extend_from_slice(&u16::MAX.to_le_bytes());
                     let end = start + u16::MAX as usize;
                     for j in start..end {
                         self.out.extend_from_slice(&self.char_buf[j]);
